Panel Config Layout Group Label add

+++ src/main/java/ngeneanalysys/controller/AnalysisDetailRawDataController.java
@@ -124,7 +124,7 @@ public class AnalysisDetailRawDataController extends AnalysisDetailCommonControl
         // 목록 컬럼 설정
         typeColumn.setCellValueFactory(cellData -> new SimpleStringProperty(cellData.getValue().getFileType()));
         filenameColumn.setCellValueFactory(cellData -> new SimpleStringProperty(cellData.getValue().getName()));
-        sizeColumn.setCellValueFactory(cellData -> new SimpleStringProperty(ConvertUtil.convertFileSizeFormat(cellData.getValue().getSize().longValue())));
+        sizeColumn.setCellValueFactory(cellData -> new SimpleStringProperty(ConvertUtil.convertFileSizeFormat(cellData.getValue().getSize())));
         createdColumn.setCellValueFactory(cellData -> new SimpleStringProperty(DateFormatUtils.format(cellData.getValue().getCreatedAt().toDate(), "yyyy/MM/dd")));
         downloadColumn.setCellValueFactory(cellData -> new SimpleObjectProperty<>(cellData.getValue()));
         downloadColumn.setCellFactory(param -> new TableCell<AnalysisFile, Object>() {
